Refs #14325: Reorder headers
