+++ ddsrouter_utils/test/unittest/memory/LesseePtrTest.cpp
@@ -20,7 +20,7 @@
 #include <gtest_aux.hpp>
 #include <gtest/gtest.h>
 
-#include <ddsrouter_utils/memory/OwnerPtr.hpp>
+#include <ddsrouter_utils/memory/owner_ptr.hpp>
 
 using namespace eprosima::ddsrouter::utils;
